+++ .gitignore
@@ -1,26 +1,35 @@
-# Allowlisting gitignore template for GO projects prevents us
-# from adding various unwanted local files, such as generated
-# files, developer configurations or IDE-specific files etc.
+# Created by https://www.toptal.com/developers/gitignore/api/go,svelte
+# Edit at https://www.toptal.com/developers/gitignore?templates=go,svelte
+
+### Go ###
+# If you prefer the allow list template instead of the deny list, see community template:
+# https://github.com/github/gitignore/blob/main/community/Golang/Go.AllowList.gitignore
 #
-# Recommended: Go.AllowList.gitignore
+# Binaries for programs and plugins
+*.exe
+*.exe~
+*.dll
+*.so
+*.dylib
 
-# Ignore everything
-*
+# Test binary, built with `go test -c`
+*.test
 
-# But not these files...
-!/.gitignore
+# Output of the go coverage tool, specifically when used with LiteIDE
+*.out
 
-!*.go
-!go.sum
-!go.mod
+# Dependency directories (remove the comment below to include it)
+# vendor/
 
-!README.md
-!LICENSE
+# Go workspace file
+go.work
 
-# !Makefile
+### Svelte ###
+# gitignore template for the SvelteKit, frontend web component framework
+# website: https://kit.svelte.dev/
 
-# ...even if they are in subdirectories
-!*/
-!Dockerfile
-!.github/workflows/*.yml
+frontend/.svelte-kit/
+frontend/package
+
+# End of https://www.toptal.com/developers/gitignore/api/go,svelte
