--- config.json
@@ -1,20 +1,0 @@
-{
-    "api_port": 8081,
-    "modbus_port": 6502,
-    "db": "./db/test.db",
-    "allow_null_register": true,
-    "registers": [
-        {
-            "tag": "210XT1055.PNT",
-            "description": "TestPoint1",
-            "address": 40001,
-            "datatype": "float32"
-        },
-        {
-            "tag": "210XT2055.PNT",
-            "description": "TestPoint2",
-            "address": 40002,
-            "datatype": "float32"
-        }
-    ]
-}
