+++ frontend/src/routes/+layout.svelte
@@ -1,0 +1,7 @@
+<nav>
+    <a href="/">Home</a>
+    <a href="/registers">Data</a>
+    <a href="/settings">Settings</a>
+</nav>
+
+<slot />
