+++ frontend/src/routes/+page.svelte
@@ -1,0 +1,1 @@
+<h1>Modbus API</h1>
