--- go.mod
@@ -1,10 +1,0 @@
-module github.com/dshargool/go-mbslave-api.git
-
-go 1.20
-
-require (
-	github.com/mattn/go-sqlite3 v1.14.17
-	github.com/simonvetter/modbus v1.6.0
-)
-
-require github.com/goburrow/serial v0.1.0 // indirect
